% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ids.R
\name{adjective_animal}
\alias{aa}
\alias{adjective_animal}
\title{Ids based on a number of adjectives and an animal}
\usage{
adjective_animal(n_adjectives = 1, style = "snake")

aa(n_adjectives = 1, style = "snake")
}
\arguments{
\item{n_adjectives}{Number of adjectives to prefix the anmial with}

\item{style}{Style to join the words with}
}
\description{
Ids based on a number of adjectives and an animal
}
\examples{
a2a <- aa(2)
a2a()
}

