% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Generic id generating function}
\usage{
ids(..., vals = list(...), style = "snake")
}
\arguments{
\item{...}{A number of character vectors}

\item{vals}{A list of character vectors, \emph{instead} of \code{...}}

\item{style}{Style to join words with}
}
\description{
Generic id generating function
}

