% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloIntegration.R
\name{plot.MCI}
\alias{plot.MCI}
\title{Plot the Monte-Carlo ingetration result}
\usage{
\method{plot}{MCI}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MCI}.}
}
\value{
A plot containing the Monte-Carlo ingetration result.
}
\description{
This function plots the Monte-Carlo ingetration result
returned by the `mc_int` function.
}
\author{
Stephane Guerrier
}
