% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_id}
\alias{random_id}
\title{Cryptographically random identifiers}
\usage{
random_id(n = 1, bytes = 16, use_openssl = TRUE)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead
returns the generating function}

\item{bytes}{The number of bytes to include for each identifier.
The length of the returned identifiers will be twice this long
with each pair of characters representing a single byte.}

\item{use_openssl}{A logical, indicating if we should use the
openssl for generating the random identifiers.  The openssl
random bytestream is not affected by the state of the R random
number generator (e.g., via \code{\link{set.seed}}) so may not be
suitable for use where reproducibility is important.  The speed
should be very similar for both approaches.}
}
\description{
Cryptopgraphically random identifiers.  This uses the
\code{openssl} package to produce a random set of bytes, and
expresses that as a hex character string.  This does not affect
R's random number stream.
}
\author{
Rich FitzJohn
}

