% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Adapted from r-bloggers
Use runShiny('rmd') to call rmd_full
Use runShiny('rmd_mini') to call rmd_mini}
\usage{
runShiny(example)
}
\description{
Adapted from r-bloggers
Use runShiny('rmd') to call rmd_full
Use runShiny('rmd_mini') to call rmd_mini
}
