% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{sentence}
\alias{sentence}
\title{Sentence style identifiers}
\usage{
sentence(n = 1, style = "lower", past = FALSE)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead
returns the generating function}

\item{style}{Style to join words with.  Can be one of "Pascal",
"camel", "snake", "kebab", "title", "sentence", "lower",
"upper", and "constant".}

\item{past}{Use the past tense for verbs (e.g., slurped or jogged
rather than slurping or jogging)}
}
\description{
Create a sentence style identifier.  This uses the approach
described by Asana on their blog
\url{https://blog.asana.com/2011/09/6-sad-squid-snuggle-softly/}
}
\author{
Rich FitzJohn
}

