% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate UUIDs}
\usage{
uuid(n = 1, drop_hyphens = FALSE)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead
returns the generating function}

\item{drop_hyphens}{Drop the hyphens from the UUID?}
}
\description{
Generate UUIDs using the uuid package
}

