% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Generic id generating function}
\usage{
ids(
  n,
  ...,
  vals = list(...),
  style = "snake",
  global = TRUE,
  use_openssl = FALSE
)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead returns
the generating function}

\item{...}{A number of character vectors}

\item{vals}{A list of character vectors, \emph{instead} of \code{...}}

\item{style}{Style to join words with.  Can be one of "Pascal",
"camel", "snake", "kebab", "dot", "title", "sentence", "lower",
"upper", "constant" or "spongemock"}

\item{global}{Use global random number generator that responds to
\code{set.seed} (see \link{random_id} for details, but
note that the default here is different).}

\item{use_openssl}{Use openssl for random number generation when
using a non-global generator (see \link{random_id} for
details)}
}
\value{
Either a character vector of length \code{n}, or a function of
one argument if \code{n} is \code{NULL}
}
\description{
Generic id generating function
}
\examples{
# For an example, please see the vignette
}
\author{
Rich FitzJohn
}
