% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Generic id generating function}
\usage{
ids(n, ..., vals = list(...), style = "snake")
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead
returns the generating function}

\item{...}{A number of character vectors}

\item{vals}{A list of character vectors, \emph{instead} of \code{...}}

\item{style}{Style to join words with.  Can be one of "Pascal",
"camel", "snake", "kebab", "dot", "title", "sentence", "lower",
"upper", and "constant".}
}
\value{
Either a character vector of length \code{n}, or a
  function of one argument if \code{n} is \code{NULL}
}
\description{
Generic id generating function
}
\examples{
# For an example, please see the vignette
}
\author{
Rich FitzJohn
}
