% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_id}
\alias{random_id}
\title{Cryptographically random identifiers}
\usage{
random_id(n = 1, bytes = 16)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead
returns the generating function}

\item{bytes}{The number of bytes to include for each identifier.
The length of the returned identifiers will be twice this long
with each pair of characters representing a single byte.}
}
\description{
Cryptopgraphically random identifiers.  This uses the
\code{openssl} package to produce a random set of bytes, and
expresses that as a hex character string.  This does not affect
R's random number stream.
}
\author{
Rich FitzJohn
}

