% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{sentence}
\alias{sentence}
\title{Sentence style identifiers}
\usage{
sentence(
  n = 1,
  style = "snake",
  past = FALSE,
  global = TRUE,
  use_openssl = NULL
)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead returns
the generating function}

\item{style}{Style to join words with.  Can be one of "Pascal",
"camel", "snake", "kebab", "dot", "title", "sentence", "lower",
"upper", "constant" or "spongemock"}

\item{past}{Use the past tense for verbs (e.g., slurped or jogged
rather than slurping or jogging)}

\item{global}{Use global random number generator that responds to
\code{set.seed} (see \link{random_id} for details, but
note that the default here is different).}

\item{use_openssl}{Use openssl for random number generation when
using a non-global generator (see \link{random_id} for
details)}
}
\description{
Create a sentence style identifier.  This uses the approach
described by Asana on their blog
\url{https://blog.asana.com/2011/09/6-sad-squid-snuggle-softly/}.
This approach encodes 32 bits of information (so 2^32 ~= 4 billion
possibilities) and in theory can be remapped to an integer if you
really wanted to.
}
\examples{
# Generate an identifier
ids::sentence()

# Generate a bunch
ids::sentence(10)

# As with adjective_animal, use "style" to control punctuation
ids::sentence(style = "Camel")
ids::sentence(style = "dot")
ids::sentence(style = "Title")

# Change the tense of the verb:
set.seed(1)
ids::sentence()
set.seed(1)
ids::sentence(past = TRUE)

# Pass n = NULL to bind arguments to a function
id <- ids::sentence(NULL, past = TRUE, style = "dot")
id()
id(10)
}
\author{
Rich FitzJohn
}
