% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate UUIDs}
\usage{
uuid(n = 1, drop_hyphens = FALSE, use_time = NA)
}
\arguments{
\item{n}{number of ids to return.  If \code{NULL}, it instead returns
the generating function}

\item{drop_hyphens}{Drop the hyphens from the UUID?}

\item{use_time}{Unused.}
}
\description{
Generate UUIDs (Universally Unique IDentifiers).  In previous
versions this was simply a thin wrapper around
\code{uuid::UUIDgenerate}, however this was subject to collisions on
windows where relatively small numbers of UUIDs generated at the
same time could return values that were identical. We now generate
only version 4 UUIDs (i.e., random though with particular bits
set).
}
\examples{
# Generate one id
ids::uuid()

# Or a bunch
ids::uuid(10)

# More in the style of random_id()
ids::uuid(drop_hyphens = TRUE)
}
\author{
Rich FitzJohn
}
